+++ advanced/kernel/src/main/java/org/neo4j/kernel/impl/nioneo/store/Store.java
@@ -32,7 +32,7 @@ public interface Store
      * @throws IOException
      *             If unable to
      */
-    public int nextId();
+    public long nextId();
     
     public long getHighestPossibleIdInUse();
